Reverted CacheableScriptFile::Seek, GetPosition, finished (for now) fixes for 64 bits.

+++ src/game/chars/CCharUse.cpp
@@ -152,8 +152,8 @@ void CChar::Use_MoonGate( CItem * pItem )
 
 		// Set it's current destination based on the moon phases.
 		// ensure iTrammelPhrase isn't smaller than iFeluccaPhase, to avoid uint underflow in next calculation
-		uint iTrammelPhase = g_World.GetMoonPhase(false) % iCount;
-		uint iFeluccaPhase = g_World.GetMoonPhase(true) % iCount;
+		size_t iTrammelPhase = g_World.GetMoonPhase(false) % iCount;
+		size_t iFeluccaPhase = g_World.GetMoonPhase(true) % iCount;
 		if ( iTrammelPhase < iFeluccaPhase )
 			iTrammelPhase += iCount;
