+++ src/sphere/strings.cpp
@@ -23,7 +23,7 @@ AbstractString::~AbstractString()
 	destroy();
 }
 
-void AbstractString::ensureLength(int newLength)
+void AbstractString::ensureLength(size_t newLength)
 {
 	// child class implements this
 	UNREFERENCED_PARAMETER(newLength);
@@ -34,12 +34,12 @@ void AbstractString::destroy()
 	// child class implements this
 }
 
-int AbstractString::length()
+size_t AbstractString::length()
 {
 	return m_length;
 }
 
-int AbstractString::realLength()
+size_t AbstractString::realLength()
 {
 	return m_realLength;
 }
@@ -54,12 +54,12 @@ const char *AbstractString::toBuffer()
 	return m_buf;
 }
 
-char AbstractString::charAt(int index)
+char AbstractString::charAt(size_t index)
 {
 	return m_buf[index];
 }
 
-void AbstractString::setAt(int index, char c)
+void AbstractString::setAt(size_t index, char c)
 {
 	m_buf[index] = c;
 }
@@ -72,7 +72,7 @@ void AbstractString::append(const char *s)
 
 void AbstractString::replace(char what, char toWhat)
 {
-	for ( int i = 0; i < m_length; i++ )
+	for (size_t i = 0; i < m_length; i++ )
 	{
 		if ( m_buf[i] == what )
 		{
@@ -123,22 +123,22 @@ bool AbstractString::startsWithHead(const char *s)
 	}
 }
 
-int AbstractString::indexOf(char c)
+size_t AbstractString::indexOf(char c)
 {
 	char *pos = strchr(m_buf, c);
-	return (int)(( pos == NULL ) ? -1 : pos - m_buf);
+	return (size_t)(( pos == NULL ) ? -1 : pos - m_buf);
 }
 
-int AbstractString::indexOf(const char *s)
+size_t AbstractString::indexOf(const char *s)
 {
 	char *pos = strstr(m_buf, s);
-	return (int)((pos == NULL) ? -1 : pos - m_buf);
+	return (size_t)((pos == NULL) ? -1 : pos - m_buf);
 }
 
-int AbstractString::lastIndexOf(char c)
+size_t AbstractString::lastIndexOf(char c)
 {
 	char *pos = strrchr(m_buf, c);
-	return (int)((pos == NULL) ? -1 : pos - m_buf);
+	return (size_t)((pos == NULL) ? -1 : pos - m_buf);
 }
 
 AbstractString::operator lpctstr() const
@@ -175,7 +175,7 @@ void String::destroy()
 	}
 }
 
-void String::ensureLength(int newLength)
+void String::ensureLength(size_t newLength)
 {
 	if ( newLength >= m_realLength )
 	{
@@ -204,7 +204,7 @@ void String::ensureLength(int newLength)
  * TemporaryString
 */
 
-int TemporaryString::m_tempPosition = 0;
+size_t TemporaryString::m_tempPosition = 0;
 char TemporaryString::m_tempStrings[MAX_TEMP_LINES_NO_CONTEXT][THREAD_STRING_LENGTH];
 
 TemporaryString::TemporaryString()
@@ -269,7 +269,7 @@ void TemporaryString::destroy()
 	}
 }
 
-void TemporaryString::ensureLength(int newLength)
+void TemporaryString::ensureLength(size_t newLength)
 {
 	if( m_useHeap )
 	{
